+++ src/main/kotlin/com/raycedni/PenAndPaperCompanion/general/stats/BaseStat.kt
@@ -2,5 +2,5 @@ package com.raycedni.PenAndPaperCompanion.general.stats
 
 interface BaseStat {
     val name: String
-    val points: Double
+    val points: Int
 }
